Fine! I'll make it long!

+++ src/main/java/com/daboxen/uberutilities/additions/materials/ExtendedFluidPipeProperties.java
@@ -8,12 +8,12 @@ public class ExtendedFluidPipeProperties extends FluidPipeProperties {
 	@Setter
 	private boolean antimatterProof;
 
-	public ExtendedFluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, boolean antimatterProof, int channels) {
+	public ExtendedFluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, boolean antimatterProof, int channels) {
 		super(maxFluidTemperature, throughput, gasProof, acidProof, cryoProof, plasmaProof, channels);
 		this.antimatterProof = antimatterProof;
 	}
 
-	public ExtendedFluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, boolean antimatterProof) {
+	public ExtendedFluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, boolean antimatterProof) {
 		this(maxFluidTemperature, throughput, gasProof, acidProof, cryoProof, plasmaProof, antimatterProof, 1);
 	}
